requirements.txt
langchain
langchain-groq
python-dotenv
mysql-connector-python
mcp
fastapi
uvicorn
pytest
